refactor: 가독성 개선을 위한 리팩터링
