+++ src/test/java/roomescape/domain/reservation/DailyThemeReservationsTest.java
@@ -1,7 +1,7 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.Assertions.assertThatCode;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -11,17 +11,13 @@ import org.junit.jupiter.api.Test;
 import roomescape.domain.member.Email;
 import roomescape.domain.member.Member;
 import roomescape.domain.member.MemberRole;
-import roomescape.domain.reservation.DailyThemeReservations;
-import roomescape.domain.reservation.Reservation;
-import roomescape.domain.reservation.ReservationTime;
-import roomescape.domain.reservation.Theme;
 import roomescape.infrastructure.error.exception.ThemeException;
 
 class DailyThemeReservationsTest {
 
     @Test
     void 생성시_같은_테마와_같은_날짜가_아닌_예약이_있으면_예외가_발생한다() {
-        //given
+        // given
         List<Reservation> reservations = List.of(
                 new Reservation(
                         1L,
@@ -39,15 +35,16 @@ class DailyThemeReservationsTest {
                 )
         );
 
-        //when & then
-        assertThatThrownBy(() -> new DailyThemeReservations(reservations, 1L, LocalDate.of(2025, 5, 8)))
+        // when
+        // then
+        assertThatCode(() -> new DailyThemeReservations(reservations, 1L, LocalDate.of(2025, 5, 8)))
                 .isInstanceOf(ThemeException.class)
                 .hasMessage("특정 테마, 특정 날짜에 속한 예약이 아닙니다.");
     }
 
     @Test
     void 예약된_시간을_계산할_수_있다() {
-        //given
+        // given
         List<Reservation> reservations = List.of(
                 new Reservation(
                         1L,
@@ -67,13 +64,16 @@ class DailyThemeReservationsTest {
         DailyThemeReservations dailyThemeReservations = new DailyThemeReservations(reservations, 1L,
                 LocalDate.of(2025, 5, 8));
 
-        //when
+        // when
         Set<ReservationTime> reservationTimes = dailyThemeReservations.calculateBookedTimes();
 
-        //then
-        assertThat(reservationTimes).isEqualTo(Set.of(
-                new ReservationTime(1L, LocalTime.of(12, 0)),
-                new ReservationTime(2L, LocalTime.of(13, 0))
-        ));
+        // then
+        assertThat(reservationTimes)
+                .isEqualTo(
+                        Set.of(
+                                new ReservationTime(1L, LocalTime.of(12, 0)),
+                                new ReservationTime(2L, LocalTime.of(13, 0))
+                        )
+                );
     }
 }
