+++ src/test/java/roomescape/domain/reservation/ReservationTimeTest.java
@@ -1,12 +1,10 @@
-package roomescape.domain;
+package roomescape.domain.reservation;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.time.LocalTime;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
-import roomescape.domain.reservation.ReservationTime;
 import roomescape.infrastructure.error.exception.ReservationTimeException;
 
 class ReservationTimeTest {
@@ -21,7 +19,7 @@ class ReservationTimeTest {
     @ParameterizedTest
     @CsvSource({"11:59", "22:01"})
     void 예약_시간은_12시_부터_22시_까지_가능하다_실패(LocalTime time) {
-        assertThatThrownBy(() -> new ReservationTime(time))
+        assertThatCode(() -> new ReservationTime(time))
                 .isInstanceOf(ReservationTimeException.class)
                 .hasMessage("해당 시간은 예약 가능 시간이 아닙니다.");
     }
